pyproject.toml
[project]
name = "ai-memory"
version = "0.1.0"
description = "AI memory system MVP with FastAPI, OpenAI, and Chroma"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0",
  "uvicorn>=0.35.0",
  "openai>=1.99.0",
  "chromadb>=1.0.16",
  "pydantic>=2.11.0",
  "python-dotenv>=1.1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4.0",
  "ruff>=0.12.0",
  "mypy>=1.17.0",
]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["api", "llm_interface", "memory_engine", "storage"]
